Time: 195 ms (24.88%), Space: 35.6 MB (27.49%) - LeetHub

+++ 1976-number-of-ways-to-arrive-at-destination/1976-number-of-ways-to-arrive-at-destination.cpp
@@ -1,5 +1,5 @@
 #define ll long long
-#define pll pair<long long, long long>
+#define pll pair<ll, ll>
 class Solution {
 public:
     int MOD = 1e9 + 7;
